citybus/city_bus_service/backend/citybus/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-js-asset==1.2.2
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django-tinymce==3.3.0
django-xhtml2pdf==0.0.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
future==0.18.2
geographiclib==1.52
geopy==2.2.0
html5lib==1.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.4.0
pycparser==2.20
PyJWT==2.3.0
PyPDF2==1.26.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2021.3
qrcode==7.3.1
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
webencodings==0.5.1
xhtml2pdf==0.2.5
